backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pyjwt==2.8.0
rocksdict==0.3.23
werkzeug==3.0.1
jsonschema==4.20.0
bcrypt==4.1.2
